test: client native 쿼리 분리
